Upgraded alphavantage to v3 framework
