+++ .changeset/great-dots-fail.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/alphavantage-adapter': major
+---
+
+Upgraded adapter from v2 to v3 framework
