--- packages/sources/alphavantage-test/CHANGELOG.md
@@ -1,33 +1,0 @@
-# @chainlink/alphavantage-test-adapter
-
-## 1.0.3
-
-### Patch Changes
-
-- cc0d39064: Bumped framework version
-- fb1b11b77: Bumped framework version
-
-## 1.0.2
-
-### Patch Changes
-
-- bfa201d6d: Bumped framework version
-- bfa201d6d: Bumped framework version
-- c600ca386: Upgrade typescript version to 5.0.4
-- bfa201d6d: Bumped framework version
-
-## 1.0.1
-
-### Patch Changes
-
-- 9bcb13c90: Bumped framework version
-
-## 1.0.0
-
-### Major Changes
-
-- 9116aa423: Add alphavantage EA v3
-
-### Patch Changes
-
-- 31af84e69: Bumped framework version
