--- packages/sources/alphavantage-test/src/endpoint/index.ts
@@ -1,1 +1,0 @@
-export { endpoint as forex } from './forex'
