--- packages/sources/alphavantage-test/test-payload.json
@@ -1,6 +1,0 @@
-{
- "requests": [{
-    "base": "GBP",
-    "quote": "USD"
- }]
-}
