--- packages/sources/alphavantage/schemas/env.json
@@ -1,22 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/alphavantage-adapter.json",
-  "title": "Chainlink External Adapter for AlphaVantage",
-  "description": "Use this adapter for connecting to [AlphaVantage's API](https://www.alphavantage.co/documentation/) from a Chainlink node.",
-  "required": ["API_KEY"],
-  "type": "object",
-  "properties": {
-    "API_ENDPOINT": {
-      "type": "string",
-      "default": "https://www.alphavantage.co/query"
-    },
-    "API_KEY": {
-      "type": "string",
-      "description": "An API key that can be obtained from [here](https://www.alphavantage.co/support/#api-key)"
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
