--- packages/sources/alphavantage/src/config/limits.json
@@ -1,24 +1,0 @@
-{
-  "http": {
-    "free": {
-      "rateLimit1m": 5,
-      "rateLimit1h": 20.83
-    },
-    "49.99": {
-      "rateLimit1m": 75
-    },
-    "99.99": {
-      "rateLimit1m": 150
-    },
-    "149.99": {
-      "rateLimit1m": 300
-    },
-    "199.99": {
-      "rateLimit1m": 600
-    },
-    "249.99": {
-      "rateLimit1m": 1200
-    }
-  },
-  "ws": {}
-}
