+++ packages/sources/alphavantage/src/endpoint/index.ts
@@ -1,5 +1,1 @@
-import type { TInputParameters as ForexInputParameters } from './forex'
-
-export type TInputParameters = ForexInputParameters
-
-export * as forex from './forex'
+export { endpoint as forex } from './forex'
