+++ packages/sources/alphavantage/test/integration/adapter.test.ts
@@ -1,46 +1,46 @@
-import { AdapterRequest } from '@chainlink/ea-bootstrap'
-import * as process from 'process'
-import { server as startServer } from '../../src'
 import { mockResponseSuccess } from './fixtures'
-import { setupExternalAdapterTest } from '@chainlink/ea-test-helpers'
-import type { SuiteContext } from '@chainlink/ea-test-helpers'
-import { SuperTest, Test } from 'supertest'
+import {
+  TestAdapter,
+  setEnvVariables,
+} from '@chainlink/external-adapter-framework/util/testing-utils'
+import * as nock from 'nock'
 
-describe('execute', () => {
-  const id = '1'
+describe('rest', () => {
+  let spy: jest.SpyInstance
+  let testAdapter: TestAdapter
+  let oldEnv: NodeJS.ProcessEnv
 
-  const context: SuiteContext = {
-    req: null,
-    server: startServer,
-  }
+  beforeAll(async () => {
+    oldEnv = JSON.parse(JSON.stringify(process.env))
+    process.env['API_KEY'] = 'fake-api-key'
+    const mockDate = new Date('2022-01-01T11:11:11.111Z')
+    spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
 
-  const envVariables = {
-    API_KEY: process.env.API_KEY || 'fake-api-key',
-    CACHE_ENABLED: 'false',
-  }
+    const adapter = (await import('./../../src')).adapter
+    adapter.rateLimiting = undefined
+    testAdapter = await TestAdapter.startWithMockedCache(adapter, {
+      testAdapter: {} as TestAdapter<never>,
+    })
+  })
 
-  setupExternalAdapterTest(envVariables, context)
+  afterAll(async () => {
+    setEnvVariables(oldEnv)
+    await testAdapter.api.close()
+    nock.restore()
+    nock.cleanAll()
+    spy.mockRestore()
+  })
 
   describe('forex rate api', () => {
-    const data: AdapterRequest = {
-      id,
-      data: {
+    it('should return success', async () => {
+      const data = {
         base: 'GBP',
         quote: 'USD',
-      },
-    }
-
-    it('should return success', async () => {
+      }
       mockResponseSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
+      const response = await testAdapter.request(data)
+      expect(response.statusCode).toBe(200)
+      expect(response.json()).toMatchSnapshot()
     })
   })
 })
