+++ packages/sources/alphavantage/test/integration/fixtures.ts
@@ -41,3 +41,4 @@ export const mockResponseSuccess = (): nock.Scope =>
         'Origin',
       ],
     )
+    .persist()
