+++ packages/tsconfig.json
@@ -138,9 +138,6 @@
       "path": "./sources/alphavantage"
     },
     {
-      "path": "./sources/alphavantage-test"
-    },
-    {
       "path": "./sources/alpine"
     },
     {
