+++ packages/tsconfig.test.json
@@ -138,9 +138,6 @@
       "path": "./sources/alphavantage/tsconfig.test.json"
     },
     {
-      "path": "./sources/alphavantage-test/tsconfig.test.json"
-    },
-    {
       "path": "./sources/alpine/tsconfig.test.json"
     },
     {
